Refactor validations test into OpenPeriodForm-tests
